add tnetstring unicode type
